Killed helpers and finished loose ends. Bot is fully functioning
