+++ comment_scan/comment_scanner.rb
@@ -1,9 +1,8 @@
 require './db'
-require_relative 'helpers'
 require_relative 'message_collection'
 
 class CommentScanner
-    attr_reader :seclient, :chatter
+    attr_reader :seclient, :chatter, :time_to_scan
 
     def initialize(seclient, chatter, post_all_comments, ignore_users, perspective_key: '', perspective_log: Logger.new('/dev/null'))
         @seclient = seclient
@@ -13,6 +12,7 @@ class CommentScanner
         @perspective_key = perspective_key
         @perspective_log = perspective_log
 
+        @time_to_scan = 60
         @latest_comment_date = @seclient.latest_comment_date.to_i+1 unless @seclient.latest_comment_date.nil?
     end
 
@@ -22,6 +22,14 @@ class CommentScanner
         scan_se_comments([new_comments])
     end
 
+    def tick
+        @time_to_scan -= 1
+        if @time_to_scan <= 0
+            @time_to_scan = 60
+            return false
+        end
+        return true
+    end
     #def scan_comment(comment_id, should_post_matches: true)
     #    #Does the work of 
     #    comment = seclient.comment_with_id(comment_id)
@@ -173,6 +181,53 @@ class CommentScanner
                 msgs.push comment, @chatter.say(report_text, room_id) if room.regex_match && report_text
             end
         end
+    end
+
+    def report_raw(post_type, comment_body)
+      regexes = Regex.where(post_type: post_type[0].downcase)
+      regexes.select do |regex|
+        %r{#{regex.regex}}.match? comment_body.downcase
+      end
+    end
+
+    def report(post_type, comment_body)
+      matching_regexes = report_raw(post_type, comment_body)
+      return "Matched regex(es) #{matching_regexes.map { |r| r.reason.nil? ? r.regex : r.reason.name }.uniq }" unless matching_regexes.empty?
+    end
+
+    def has_magic_comment?(comment, post)
+      !comment.body_markdown.include?("https://interpersonal.meta.stackexchange.com/q/1644/31") &&
+      post.comments.any? do |c|
+        c.body_markdown.include?("https://interpersonal.meta.stackexchange.com/q/1644/31")
+      end
+    end
+
+    def ts_for(ts)
+      return "" if ts.nil?
+      ts = (Time.new - Time.at(ts.to_i)).to_i
+      return "" if ts < 0
+      if ts < 60
+        "#{ts} seconds ago"
+      elsif ts/60 < 60
+        "#{ts/60} minutes ago"
+      elsif ts/(60**2) < 60
+        "#{ts/(60**2)} hours ago"
+      else
+        "#{ts/(24*60*60)} days ago"
+      end
+    end
+
+    def timestamp_to_date(timestamp)
+      Time.at(timestamp.to_i).to_date
+    end
+
+    def user_for(author)
+      #return "" unless author.is_a? SE::API::User
+      name = author.name
+      link = author.link&.gsub(/(^.*u[sers]{4}?\/\d*)\/.*$/, '\1')&.gsub("/users/", "/u/")
+      rep = author.reputation
+      return "(deleted user)" if name.nil? && link.nil? && rep.nil?
+      "[#{name}](#{link}) (#{rep} rep)"
     end
 
     def perspective_scan(text)
